:pushpin: Added automatic version capture

+++ src/main/java/com/yuhtin/lauren/events/ChatMessage.java
@@ -8,6 +8,8 @@ import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
 
+import java.util.concurrent.TimeUnit;
+
 public class ChatMessage extends ListenerAdapter {
 
     /* Earn 3 XP for every message sent */
@@ -21,7 +23,7 @@ public class ChatMessage extends ListenerAdapter {
             String command = event.getMessage().getContentRaw().split(" ")[0].replace("$", "");
             if (!CommandCache.aliases.contains(command)) {
                 for (String alias : CommandCache.aliases) {
-                    if (LevenshteinCalculator.eval(command, alias) < 6) {
+                    if (LevenshteinCalculator.eval(command, alias) < 4) {
                         event.getChannel().sendMessage("<:chorano:726207542413230142> Esse comandinho não existe porém encontrei um parecido: `$" + alias + "`").queue();
                         break;
                     }
@@ -29,12 +31,19 @@ public class ChatMessage extends ListenerAdapter {
             }
         }
 
-        if (event.getMessage().getContentRaw().contains("https://") && !Utilities.isPermission(event.getMember(), event.getChannel(), Permission.MESSAGE_MANAGE, false)) {
-            event.getChannel().sendMessage("<:chorano:726207542413230142> Poxa, não divulga aqui amigo, temos nosso sistema de parceria, fale com o <@272879983326658570> no privado.").queue();
+        if (!isMusicCommand(event.getMessage().getContentRaw()) && event.getMessage().getContentRaw().contains("https://") && !Utilities.isPermission(event.getMember(), event.getChannel(), Permission.MESSAGE_MANAGE, false)) {
+            event.getChannel().sendMessage("<:chorano:726207542413230142> Poxa, não divulga aqui amigo, temos nosso sistema de parceria, fale com o <@272879983326658570> no privado.")
+                    .queue(m -> m.delete().queueAfter(5, TimeUnit.SECONDS));
             event.getMessage().delete().queue();
             return;
         }
 
         PlayerDataController.get(event.getMember().getIdLong()).gainXP(3).updateLevel().save();
     }
+
+    private boolean isMusicCommand(String contentRaw) {
+        return contentRaw.startsWith("$m ") || contentRaw.startsWith("$music") || contentRaw.startsWith("$play ") || contentRaw.startsWith("$tocar ");
+    }
+
+
 }
